requirements.txt
APScheduler==3.6.3
asgiref==3.8.1
backoff==2.2.1
cachetools==4.2.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
decorator==5.1.1
Django==5.0.6
djangorestframework==3.15.2
dnspython==2.6.1
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
gspread==6.1.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
importlib-resources==6.4.0
monotonic==1.6
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
ploomber-cloud==0.2.14
ploomber-core==0.2.25
posthog==3.5.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pymongo==4.7.3
pyparsing==3.1.2
pyshorteners==1.0.1
pyTelegramBotAPI==4.20.0
python-dateutil==2.9.0.post0
python-telegram-bot
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
validators==0.20.0
